+++ lib/components/Textarea/Textarea.d.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+export interface ITextareaProps extends React.DetailedHTMLProps<React.TextareaHTMLAttributes<HTMLTextAreaElement>, HTMLTextAreaElement> {
+}
+export declare const Textarea: (props: ITextareaProps) => JSX.Element;