+++ src/lib/components/index.ts
@@ -19,6 +19,7 @@ export * from './SvgIcon';
 export * from './Snackbar';
 export * from './Table';
 export * from './Tabset';
+export * from './Textarea';
 export * from './Tooltip';
 export * from './TreeView';
 export * from './component.enums';